refactor: 검증 클래스를 생성해 검증 메서드를 분리합니다.
