+++ src/main/java/calculator/util/Calculate.java
@@ -6,8 +6,6 @@ import java.util.List;
 
 public class Calculate {
 
-    private static final Integer ZERO = 0;
-
     public static Integer calculateUserInputSum(String userInput) {
         return calculateSplitUserInputSum(calculateSplitUserInput(userInput));
     }
@@ -18,13 +16,5 @@ public class Calculate {
 
     public static Integer calculateSplitUserInputSum(List<String> userInput) {
         return new Numbers(userInput).calculateNumbersSum();
-    }
-
-    public static boolean checkNotContainAllSeparator(String userInput) {
-        return CalculateDefault.checkNotContainDefault(userInput) && CalculateCustom.checkNotContainCustom(userInput);
-    }
-
-    public static boolean checkIsNumberPositive(Integer number){
-        return number < ZERO;
     }
 }
