+++ src/main/java/calculator/util/CalculatorValidation.java
@@ -1,0 +1,14 @@
+package calculator.util;
+
+public class CalculatorValidation {
+
+    private static final Integer ZERO = 0;
+
+    public static boolean checkNotContainAllSeparator(String userInput) {
+        return CalculateDefault.checkNotContainDefault(userInput) && CalculateCustom.checkNotContainCustom(userInput);
+    }
+
+    public static boolean checkIsNumberPositive(Integer number){
+        return number < ZERO;
+    }
+}
